requirements.txt
fastapi[all]
uvicorn
langchain
langgraph
langchain-openai
instructor
pydantic
pydantic-settings
sqlalchemy
sqlmodel
psycopg2-binary
chromadb
sentence-transformers
pdfplumber
python-docx
redis
celery
python-multipart
python-dotenv
pytest
httpx
